update other example
